+++ src/main/java/org/betonquest/betonquest/compatibility/citizens/objective/NPCKillObjective.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.compatibility.citizens;
+package org.betonquest.betonquest.compatibility.citizens.objective;
 
 import net.citizensnpcs.api.CitizensAPI;
 import net.citizensnpcs.api.npc.NPC;
@@ -17,11 +17,18 @@ import org.bukkit.event.Listener;
 /**
  * Player has to kill an NPC.
  */
-@SuppressWarnings("PMD.CommentRequired")
 public class NPCKillObjective extends CountingObjective implements Listener {
-
+    /**
+     * The NPC id.
+     */
     private final int npcId;
 
+    /**
+     * Create a new Citizens NPC kill objective.
+     *
+     * @param instruction the user-provided instruction
+     * @throws InstructionParseException when the instruction cannot be parsed or is invalid
+     */
     public NPCKillObjective(final Instruction instruction) throws InstructionParseException {
         super(instruction, "mobs_to_kill");
         npcId = instruction.getInt();
@@ -31,13 +38,15 @@ public class NPCKillObjective extends CountingObjective implements Listener {
         targetAmount = instruction.getVarNum(instruction.getOptional("amount", "1"), VariableNumber.NOT_LESS_THAN_ONE_CHECKER);
     }
 
+    /**
+     * Handles the NPC death.
+     *
+     * @param event the event to listen
+     */
     @EventHandler(ignoreCancelled = true)
     public void onNpcKill(final MobKilledEvent event) {
         final NPC npc = CitizensAPI.getNPCRegistry().getNPC(event.getEntity());
-        if (npc == null) {
-            return;
-        }
-        if (npc.getId() != npcId) {
+        if (npc == null || npc.getId() != npcId) {
             return;
         }
         final OnlineProfile onlineProfile = event.getProfile().getOnlineProfile().get();
